feat: refactoring detail page ditributing
- distribute events and html
- remove console, don't needed
