+++ frontend/src/html/detail.html
@@ -1,22 +1,3 @@
-<header class="header">
-  <h1 class="header--logo">develog</h1>
-  <form class="search--form" action="">
-    <input id="search" class="search--hidden" type="text" />
-    <label for="search" class="fas fa-search"></label>
-  </form>
-  <button class="button button--login">Login</button>
-  <div class="user hidden">
-    <img class="avatar" src="#" alt="avatar image" />
-  </div>
-  <nav class="nav-box hidden">
-    <ul>
-      <li>내 블로그</li>
-      <li>마이페이지</li>
-      <li>로그아웃</li>
-    </ul>
-  </nav>
-</header>
-
 <main class="detail-container">
   <section class="detail">
     <h1 class="detail__title"></h1>
@@ -25,13 +6,16 @@
         <img class="avatar-button avatar-button--size" src="#" alt="avatar-button" />
       </button>
       <span class="author"></span>
-      <button class="no-display heart-btn btn">
+      <button class="heart-btn btn">
         <i class="far fa-heart"></i>
       </button>
-      <button class="none edit pencil-btn btn">
+      <button class="fullheart heart-btn btn">
+        <i class="fas fa-heart"></i>
+      </button>
+      <button class="edit pencil-btn btn">
         <i class="bx bx-pencil"></i>
       </button>
-      <button class="none edit trash-btn btn">
+      <button class="edit trash-btn btn">
         <i class="far fa-trash-alt"></i>
       </button>
     </div>
@@ -42,7 +26,7 @@
   <section class="user-comment">
     <div class="comment__user-info">
       <button class="btn">
-        <img class="avatar-button avatar-button--size" src="#" class="user__avatar" alt="user-avatar" />
+        <img class="avatar-button avatar-button--size" src="#" alt="user-avatar" />
       </button>
       <span class="user-id"></span>
     </div>
@@ -59,7 +43,7 @@
     <div class="comment">
       <div class="comment__user-info">
         <button class="btn">
-          <img class="user__avatar avatar-button avatar-button--size" src="#" alt="comments-avatar" />
+          <img class="avatar-button avatar-button--size" src="#" alt="user-avatar" />
         </button>
         <span class="user-id">minsoftk</span>
       </div>
