+++ frontend/src/js/app.js
@@ -3,6 +3,8 @@ import Main from './main';
 
 // import detailUrlEvents from './js/detail';
 import Signin from './signin';
+import Detail from './detail';
+import Write from './write';
 // import signup from './js/signup';
 // import develog from './js/develog';
 // import mypage from './js/mypage';
@@ -14,14 +16,14 @@ import Signin from './signin';
 const routes = [
   // { path: '/', component: render(mainNode) },
   { path: '/', component: Main },
+  { path: '/detail', component: Detail },
+  { path: '/write', component: Write },
+  { path: '/signin', component: Signin },
   // { path: '/search', component: Search },
   // { path: '/develog', component: Develog },
-  // { path: '/detail', component: Detail },
-  { path: '/signin', component: Signin },
   // { path: '/signup', component: Signup },
   // { path: '/mypage', component: Mypage },
   // { path: '/mypageEdit', component: MypageEdit },
-  // { path: '/write', component: Write },
   // { path: '**', component: NotFound },
 ];
 
@@ -29,7 +31,7 @@ const router = () => {
   const path = window.location.pathname.split('/');
   const targetPath = '/' + path[1];
 
-  return [, ...routes.find(route => route.path === targetPath).component()];
+  return routes.find(route => route.path === targetPath).component();
 };
 
 export default router;
